Use the default cargo format
