--- cli/rustfmt.toml
@@ -1,23 +1,0 @@
-# Basic
-hard_tabs = true
-max_width = 100
-use_small_heuristics = "Max"
-
-# Imports
-imports_granularity = "Crate"
-reorder_imports = true
-
-# Consistency
-newline_style = "Unix"
-
-# Misc
-binop_separator = "Back"
-chain_width = 80
-match_arm_blocks = false
-match_arm_leading_pipes = "Preserve"
-match_block_trailing_comma = true
-reorder_impl_items = false
-spaces_around_ranges = false
-trailing_comma = "Vertical"
-trailing_semicolon = false
-use_field_init_shorthand = true
