+++ cli/src/apis/configuration.rs
@@ -12,23 +12,23 @@
 
 #[derive(Debug, Clone)]
 pub struct Configuration {
-	pub base_path: String,
-	pub user_agent: Option<String>,
-	pub client: reqwest::Client,
+    pub base_path: String,
+    pub user_agent: Option<String>,
+    pub client: reqwest::Client,
 }
 
 impl Configuration {
-	pub fn new() -> Configuration {
-		Configuration::default()
-	}
+    pub fn new() -> Configuration {
+        Configuration::default()
+    }
 }
 
 impl Default for Configuration {
-	fn default() -> Self {
-		Configuration {
-			base_path: "http://0.0.0.0:8000".to_owned(),
-			user_agent: Some("Agent-CLI/v1".to_owned()),
-			client: reqwest::Client::new(),
-		}
-	}
+    fn default() -> Self {
+        Configuration {
+            base_path: "http://0.0.0.0:8000".to_owned(),
+            user_agent: Some("Agent-CLI/v1".to_owned()),
+            client: reqwest::Client::new(),
+        }
+    }
 }
