+++ cli/src/models/artifact.rs
@@ -12,28 +12,33 @@
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Artifact {
-	/// ID of the artifact.
-	#[serde(rename = "artifact_id")]
-	pub artifact_id: String,
-	/// Whether the artifact has been created by the agent.
-	#[serde(rename = "agent_created")]
-	pub agent_created: bool,
-	/// Filename of the artifact.
-	#[serde(rename = "file_name")]
-	pub file_name: String,
-	/// Relative path of the artifact in the agent's workspace.
-	#[serde(
-		rename = "relative_path",
-		default,
-		with = "::serde_with::rust::double_option",
-		skip_serializing_if = "Option::is_none"
-	)]
-	pub relative_path: Option<Option<String>>,
+    /// ID of the artifact.
+    #[serde(rename = "artifact_id")]
+    pub artifact_id: String,
+    /// Whether the artifact has been created by the agent.
+    #[serde(rename = "agent_created")]
+    pub agent_created: bool,
+    /// Filename of the artifact.
+    #[serde(rename = "file_name")]
+    pub file_name: String,
+    /// Relative path of the artifact in the agent's workspace.
+    #[serde(
+        rename = "relative_path",
+        default,
+        with = "::serde_with::rust::double_option",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub relative_path: Option<Option<String>>,
 }
 
 impl Artifact {
-	/// An Artifact either created by or submitted to the agent.
-	pub fn new(artifact_id: String, agent_created: bool, file_name: String) -> Artifact {
-		Artifact { artifact_id, agent_created, file_name, relative_path: None }
-	}
+    /// An Artifact either created by or submitted to the agent.
+    pub fn new(artifact_id: String, agent_created: bool, file_name: String) -> Artifact {
+        Artifact {
+            artifact_id,
+            agent_created,
+            file_name,
+            relative_path: None,
+        }
+    }
 }
