+++ cli/src/models/pagination.rs
@@ -10,27 +10,32 @@
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct Pagination {
-	/// Total number of items.
-	#[serde(rename = "total_items")]
-	pub total_items: i32,
-	/// Total number of pages.
-	#[serde(rename = "total_pages")]
-	pub total_pages: i32,
-	/// Current_page page number.
-	#[serde(rename = "current_page")]
-	pub current_page: i32,
-	/// Number of items per page.
-	#[serde(rename = "page_size")]
-	pub page_size: i32,
+    /// Total number of items.
+    #[serde(rename = "total_items")]
+    pub total_items: i32,
+    /// Total number of pages.
+    #[serde(rename = "total_pages")]
+    pub total_pages: i32,
+    /// Current_page page number.
+    #[serde(rename = "current_page")]
+    pub current_page: i32,
+    /// Number of items per page.
+    #[serde(rename = "page_size")]
+    pub page_size: i32,
 }
 
 impl Pagination {
-	pub fn new(
-		total_items: i32,
-		total_pages: i32,
-		current_page: i32,
-		page_size: i32,
-	) -> Pagination {
-		Pagination { total_items, total_pages, current_page, page_size }
-	}
+    pub fn new(
+        total_items: i32,
+        total_pages: i32,
+        current_page: i32,
+        page_size: i32,
+    ) -> Pagination {
+        Pagination {
+            total_items,
+            total_pages,
+            current_page,
+            page_size,
+        }
+    }
 }
