+++ cli/src/models/step_request_body.rs
@@ -12,22 +12,25 @@
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct StepRequestBody {
-	/// Input prompt for the step.
-	#[serde(
-		rename = "input",
-		default,
-		with = "::serde_with::rust::double_option",
-		skip_serializing_if = "Option::is_none"
-	)]
-	pub input: Option<Option<String>>,
-	/// Input parameters for the task step. Any value is allowed.
-	#[serde(rename = "additional_input", skip_serializing_if = "Option::is_none")]
-	pub additional_input: Option<serde_json::Value>,
+    /// Input prompt for the step.
+    #[serde(
+        rename = "input",
+        default,
+        with = "::serde_with::rust::double_option",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub input: Option<Option<String>>,
+    /// Input parameters for the task step. Any value is allowed.
+    #[serde(rename = "additional_input", skip_serializing_if = "Option::is_none")]
+    pub additional_input: Option<serde_json::Value>,
 }
 
 impl StepRequestBody {
-	/// Body of the task request.
-	pub fn new() -> StepRequestBody {
-		StepRequestBody { input: None, additional_input: None }
-	}
+    /// Body of the task request.
+    pub fn new() -> StepRequestBody {
+        StepRequestBody {
+            input: None,
+            additional_input: None,
+        }
+    }
 }
