+++ cli/src/models/task_request_body.rs
@@ -12,22 +12,25 @@
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct TaskRequestBody {
-	/// Input prompt for the task.
-	#[serde(
-		rename = "input",
-		default,
-		with = "::serde_with::rust::double_option",
-		skip_serializing_if = "Option::is_none"
-	)]
-	pub input: Option<Option<String>>,
-	/// Input parameters for the task. Any value is allowed.
-	#[serde(rename = "additional_input", skip_serializing_if = "Option::is_none")]
-	pub additional_input: Option<serde_json::Value>,
+    /// Input prompt for the task.
+    #[serde(
+        rename = "input",
+        default,
+        with = "::serde_with::rust::double_option",
+        skip_serializing_if = "Option::is_none"
+    )]
+    pub input: Option<Option<String>>,
+    /// Input parameters for the task. Any value is allowed.
+    #[serde(rename = "additional_input", skip_serializing_if = "Option::is_none")]
+    pub additional_input: Option<serde_json::Value>,
 }
 
 impl TaskRequestBody {
-	/// Body of the task request.
-	pub fn new() -> TaskRequestBody {
-		TaskRequestBody { input: None, additional_input: None }
-	}
+    /// Body of the task request.
+    pub fn new() -> TaskRequestBody {
+        TaskRequestBody {
+            input: None,
+            additional_input: None,
+        }
+    }
 }
