+++ cli/src/models/task_steps_list_response.rs
@@ -10,17 +10,20 @@
 
 #[derive(Clone, Debug, PartialEq, Serialize, Deserialize)]
 pub struct TaskStepsListResponse {
-	#[serde(rename = "steps")]
-	pub steps: Vec<crate::models::Step>,
-	#[serde(rename = "pagination")]
-	pub pagination: Box<crate::models::Pagination>,
+    #[serde(rename = "steps")]
+    pub steps: Vec<crate::models::Step>,
+    #[serde(rename = "pagination")]
+    pub pagination: Box<crate::models::Pagination>,
 }
 
 impl TaskStepsListResponse {
-	pub fn new(
-		steps: Vec<crate::models::Step>,
-		pagination: crate::models::Pagination,
-	) -> TaskStepsListResponse {
-		TaskStepsListResponse { steps, pagination: Box::new(pagination) }
-	}
+    pub fn new(
+        steps: Vec<crate::models::Step>,
+        pagination: crate::models::Pagination,
+    ) -> TaskStepsListResponse {
+        TaskStepsListResponse {
+            steps,
+            pagination: Box::new(pagination),
+        }
+    }
 }
